+++ Business/Abstract/ICarService.cs
@@ -11,9 +11,9 @@ namespace Business.Abstract
     {
         IDataResult<List<Car>> GetAll();
 
-        IDataResult<List<Car>> GetCarsByBrandId(int brandId);
+        IDataResult<List<Car>> GetCarsByBrandId(long brandId);
 
-        IDataResult<List<Car>> GetCarsByColorId(int colorId);
+        IDataResult<List<Car>> GetCarsByColorId(long colorId);
 
         IResult Add(Car car);
