+++ Entities/Concrete/Car.cs
@@ -8,12 +8,12 @@ namespace Entities.Concrete
 {
     public class Car : IEntity
     {
-        public int Id { get; set; }
-        public int BrandId { get; set; }
-        public int ColorId { get; set; }
-        public int ModelYear { get; set; }
+        public long Id { get; set; }
+        public long BrandId { get; set; }
+        public long ColorId { get; set; }
+        public long ModelYear { get; set; }
 
-        [Range(1, int.MaxValue, ErrorMessage = "DailyPrice should be bigger than 0")]
+        [Range(1, long.MaxValue, ErrorMessage = "DailyPrice should be bigger than 0")]
         public decimal DailyPrice { get; set; }
 
         [MinLength(2,ErrorMessage ="Car Name Should Be min 2 character long")]
